Restructure search-ui components (move towards low level building blocks for the search sidebar panel)
