+++ client/search-ui/src/index.ts
@@ -13,6 +13,10 @@ export * from './results/aggregation'
 export * from './results/sidebar/FilterLink'
 export * from './results/sidebar/revisions'
 export * from './results/sidebar/SearchSidebar'
+export * from './results/sidebar/QuickLink'
+export * from './results/sidebar/helpers'
+export * from './results/sidebar/SearchReference'
+export * from './results/sidebar/SearchTypeLink'
 export * from './results/sidebar/SidebarButtonStrip'
 export * from './results/StreamingSearchResultsList'
 export * from './util'
