+++ client/search-ui/src/results/sidebar/SearchFilterSection.module.scss
