+++ client/search-ui/src/results/sidebar/SearchFilterSection.test.tsx
@@ -5,7 +5,7 @@ import sinon from 'sinon'
 import { Filter } from '@sourcegraph/shared/src/search/stream'
 
 import { getDynamicFilterLinks } from './FilterLink'
-import { SearchSidebarSection } from './SearchSidebarSection'
+import { SearchFilterSection } from './SearchFilterSection'
 
 describe('SearchSidebarSection', () => {
     const filters: Filter[] = ['typescript', 'JavaScript', 'c++', 'c', 'c#', 'python', 'ruby', 'haskell', 'java'].map(
@@ -22,21 +22,20 @@ describe('SearchSidebarSection', () => {
 
     it('should render all items initially', () => {
         render(
-            <SearchSidebarSection sectionId="id" header="Dynamic filters" showSearch={true}>
+            <SearchFilterSection sectionId="id" header="Dynamic filters" showSearch={true}>
                 {getDynamicFilterLinks(filters, ['file', 'lang', 'utility'], onFilterChosen)}
-            </SearchSidebarSection>
+            </SearchFilterSection>
         )
 
         expect(screen.getAllByTestId('filter-link')).toHaveLength(9)
-
         expect(screen.getByTestId('sidebar-section-search-box')).toBeInTheDocument()
     })
 
     it('should filter items based on search', () => {
         render(
-            <SearchSidebarSection sectionId="id" header="Dynamic filters" showSearch={true}>
+            <SearchFilterSection sectionId="id" header="Dynamic filters" showSearch={true}>
                 {getDynamicFilterLinks(filters, ['file', 'lang', 'utility'], onFilterChosen)}
-            </SearchSidebarSection>
+            </SearchFilterSection>
         )
 
         userEvent.type(screen.getByTestId('sidebar-section-search-box'), 'Script')
@@ -46,49 +45,46 @@ describe('SearchSidebarSection', () => {
 
     it('should clear search when items change', () => {
         const { rerender } = render(
-            <SearchSidebarSection sectionId="id" header="Dynamic filters" showSearch={true}>
+            <SearchFilterSection sectionId="id" header="Dynamic filters" showSearch={true}>
                 {getDynamicFilterLinks(filters, ['file', 'lang', 'utility'], onFilterChosen)}
-            </SearchSidebarSection>
+            </SearchFilterSection>
         )
 
         userEvent.type(screen.getByTestId('sidebar-section-search-box'), 'Script')
 
         rerender(
-            <SearchSidebarSection sectionId="id" header="Dynamic filters" showSearch={true}>
+            <SearchFilterSection sectionId="id" header="Dynamic filters" showSearch={true}>
                 {getDynamicFilterLinks(
                     [filters[0], filters[5], filters[3]],
                     ['file', 'lang', 'utility'],
                     onFilterChosen
                 )}
-            </SearchSidebarSection>
+            </SearchFilterSection>
         )
 
         expect(screen.getAllByTestId('filter-link')).toHaveLength(3)
-
         expect(screen.getByTestId('sidebar-section-search-box')).toHaveValue('')
     })
 
     it('should not show search if only one item in list', () => {
         render(
-            <SearchSidebarSection sectionId="id" header="Dynamic filters" showSearch={true}>
+            <SearchFilterSection sectionId="id" header="Dynamic filters" showSearch={true}>
                 {getDynamicFilterLinks([filters[2]], ['file', 'lang', 'utility'], onFilterChosen)}
-            </SearchSidebarSection>
+            </SearchFilterSection>
         )
 
         expect(screen.getByTestId('filter-link')).toBeInTheDocument()
-
         expect(screen.queryByTestId('sidebar-section-search-box')).not.toBeInTheDocument()
     })
 
     it('should not show search if showSearch is false', () => {
         render(
-            <SearchSidebarSection sectionId="id" header="Dynamic filters">
+            <SearchFilterSection sectionId="id" header="Dynamic filters">
                 {getDynamicFilterLinks(filters, ['file', 'lang', 'utility'], onFilterChosen)}
-            </SearchSidebarSection>
+            </SearchFilterSection>
         )
 
         expect(screen.getAllByTestId('filter-link')).toHaveLength(9)
-
         expect(screen.queryByTestId('sidebar-section-search-box')).not.toBeInTheDocument()
     })
 })
